Fixed error call to undefined function `Ninja_Forms()`
Problem was the hook which added post meta to the submission: previously it was
`wp_loaded` with priority 5, but the function didn't exist then (or rather, it
used to in an old version of Ninja Forms but now doesn't).

Moving the action to the `wp` hook with priority 11 resolved the issue.